nz-energy-grid/requirements.txt
flask>=3.0
requests>=2.31
python-dotenv>=1.0
gunicorn>=21.2
pytz>=2024.1
